sensor/serializers.py
from rest_framework import serializers
from sensor.models import SensorData


class SensorDataSerializer(serializers.ModelSerializer):
    class Meta:
        model = SensorData
        fields = '__all__'
